Renaming after code review
