+++ src/app/routes/onDemandRadio/getInitialData/index.test.js
@@ -4,7 +4,7 @@ import onDemandRadioJson from '#data/pashto/bbc_pashto_radio/w172x8nvf4bchz5.jso
 fetch.mockResponse(JSON.stringify(onDemandRadioJson));
 
 it('should return essential data for a page to render', async () => {
-  const { pageData } = await getInitialData('mock-live-radio-path');
+  const { pageData } = await getInitialData('mock-on-demand-radio-path');
 
   expect(pageData.promo.headlines.headline).toEqual('وروستي خبرونه');
   expect(pageData.metadata.id).toEqual(
